[#80] * moved factories to separate package
