+++ gradoop-common/src/test/java/org/gradoop/common/model/impl/pojo/VertexTest.java
@@ -16,6 +16,7 @@
 package org.gradoop.common.model.impl.pojo;
 
 import org.gradoop.common.model.api.entities.Vertex;
+import org.gradoop.common.model.impl.factories.EPGMVertexFactory;
 import org.gradoop.common.model.impl.id.GradoopId;
 import org.gradoop.common.model.impl.id.GradoopIdSet;
 import org.gradoop.common.model.impl.properties.Properties;
