+++ gradoop-data-integration/src/test/java/org/gradoop/dataintegration/importer/impl/json/functions/MinimalJsonToVertexTest.java
@@ -18,7 +18,7 @@ package org.gradoop.dataintegration.importer.impl.json.functions;
 import org.codehaus.jettison.json.JSONException;
 import org.gradoop.common.model.api.entities.Vertex;
 import org.gradoop.common.model.impl.pojo.EPGMVertex;
-import org.gradoop.common.model.impl.pojo.EPGMVertexFactory;
+import org.gradoop.common.model.impl.factories.EPGMVertexFactory;
 import org.gradoop.common.model.impl.properties.PropertyValue;
 import org.gradoop.flink.model.GradoopFlinkTestBase;
 import org.junit.Before;
