+++ gradoop-data-integration/src/test/java/org/gradoop/dataintegration/transformation/functions/CreateVertexFromEdgesTest.java
@@ -19,7 +19,7 @@ import org.apache.flink.api.java.tuple.Tuple3;
 import org.gradoop.common.model.impl.id.GradoopId;
 import org.gradoop.common.model.impl.pojo.EPGMEdge;
 import org.gradoop.common.model.impl.pojo.EPGMVertex;
-import org.gradoop.common.model.impl.pojo.EPGMEdgeFactory;
+import org.gradoop.common.model.impl.factories.EPGMEdgeFactory;
 import org.gradoop.flink.model.GradoopFlinkTestBase;
 import org.junit.Test;
