+++ gradoop-flink/src/test/java/org/gradoop/flink/model/impl/functions/filters/AbstractRichCombinedFilterFunctionTest.java
@@ -20,7 +20,7 @@ import org.apache.flink.api.java.DataSet;
 import org.apache.flink.api.java.io.LocalCollectionOutputFormat;
 import org.apache.flink.configuration.Configuration;
 import org.gradoop.common.model.impl.pojo.EPGMVertex;
-import org.gradoop.common.model.impl.pojo.EPGMVertexFactory;
+import org.gradoop.common.model.impl.factories.EPGMVertexFactory;
 import org.gradoop.flink.model.GradoopFlinkTestBase;
 import org.gradoop.flink.model.impl.functions.epgm.IdInBroadcast;
 import org.junit.Test;
