+++ gradoop-flink/src/test/java/org/gradoop/flink/model/impl/operators/matching/common/query/predicates/CNFTest.java
@@ -18,7 +18,7 @@ package org.gradoop.flink.model.impl.operators.matching.common.query.predicates;
 import com.google.common.collect.Sets;
 import org.gradoop.common.model.impl.id.GradoopId;
 import org.gradoop.common.model.impl.pojo.EPGMVertex;
-import org.gradoop.common.model.impl.pojo.EPGMVertexFactory;
+import org.gradoop.common.model.impl.factories.EPGMVertexFactory;
 import org.gradoop.common.model.impl.properties.Properties;
 import org.gradoop.common.model.impl.properties.PropertyValue;
 import org.gradoop.flink.model.impl.operators.matching.common.query.QueryHandler;
