+++ gradoop-store/gradoop-accumulo/src/main/java/org/gradoop/storage/impl/accumulo/io/inputformats/VertexInputFormat.java
@@ -23,7 +23,7 @@ import org.apache.accumulo.core.data.Key;
 import org.apache.accumulo.core.data.Value;
 import org.gradoop.common.model.api.entities.Vertex;
 import org.gradoop.common.model.impl.pojo.EPGMVertex;
-import org.gradoop.common.model.impl.pojo.EPGMVertexFactory;
+import org.gradoop.common.model.impl.factories.EPGMVertexFactory;
 import org.gradoop.storage.impl.accumulo.constants.AccumuloTables;
 import org.gradoop.storage.impl.accumulo.handler.AccumuloVertexHandler;
 import org.gradoop.storage.impl.accumulo.iterator.tserver.GradoopVertexIterator;
