+++ gradoop-store/gradoop-accumulo/src/main/java/org/gradoop/storage/impl/accumulo/iterator/tserver/GradoopGraphHeadIterator.java
@@ -22,7 +22,7 @@ import org.apache.accumulo.core.util.Pair;
 import org.gradoop.common.model.api.entities.GraphHead;
 import org.gradoop.common.model.impl.id.GradoopId;
 import org.gradoop.common.model.impl.pojo.EPGMGraphHead;
-import org.gradoop.common.model.impl.pojo.EPGMGraphHeadFactory;
+import org.gradoop.common.model.impl.factories.EPGMGraphHeadFactory;
 import org.gradoop.common.model.impl.properties.PropertyValue;
 import org.gradoop.storage.impl.accumulo.constants.AccumuloTables;
 import org.gradoop.storage.utils.KryoUtils;
